chore: move event replay functions to separate file
